backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
greenlet



sqlalchemy[asyncio]

pydantic[email]



openai
authlib
itsdangerous


motor
PyMuPDF

pydantic>=2.0
pydantic-settings

python-multipart
python-dotenv
numpy

passlib[bcrypt]

python-jose[cryptography]


pytest
httpx
